+++ src/redturtle/prenotazioni/content/prenotazioni_folder.py
@@ -15,12 +15,14 @@ from z3c.form import validator
 from z3c.form.browser.checkbox import CheckBoxFieldWidget
 from zope import schema
 from zope.component import provideAdapter
+from zope.interface import provider
 from zope.interface import implementer
 from zope.interface import Interface
 from zope.interface import Invalid
 from zope.interface import invariant
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
+from zope.schema.interfaces import IContextAwareDefaultFactory
 
 try:
     from plone.app.dexterity import textindexer
@@ -131,6 +133,95 @@ class IBookingTypeRow(Interface):
         title=_("Duration value"),
         required=True,
         vocabulary="redturtle.prenotazioni.VocDurataIncontro",
+    )
+
+
+@provider(IContextAwareDefaultFactory)
+def templates_usage_default_factory(context):
+    return _(
+        "templates_usage_default_value",
+        "<ul>"
+        "<li>${title} - title.</li>"
+        "<li>${booking_gate} - booking gate.</li>"
+        "<li>${booking_human_readable_start} - booking human readable start.</li>"
+        "<li>${booking_date} - booking date.</li>"
+        "<li>${booking_end_date} - booking end date.</li>"
+        "<li>${booking_time} - booking time.</li>"
+        "<li>${booking_time_end} - booking time end.</li>"
+        "<li>${booking_code} - booking code.</li>"
+        "<li>${booking_type} - booking type.</li>"
+        "<li>${booking_print_url} - booking print url.</li>"
+        "<li>${booking_url_with_delete_token} - booking url with delete token.</li>"
+        "<li>${booking_user_phone} - booking user phone.</li>"
+        "<li>${booking_user_email} - booking user email.</li>"
+        "<li>${booking_office_contact_phone} - booking office contact phone.</li>"
+        "<li>${booking_office_contact_pec} - booking office contact pec.</li>"
+        "<li>${booking_office_contact_fax} - booking office contact fax.</li>"
+        "<li>${booking_how_to_get_to_office} - booking how to get to office.</li>"
+        "<li>${booking_office_complete_address} - booking office complete address.</li>"
+        "</ul>",
+    )
+
+
+@provider(IContextAwareDefaultFactory)
+def notify_on_submit_subject_default_factory(context):
+    return _("notify_on_submit_subject_default_value", "Booking created ${title}")
+
+
+@provider(IContextAwareDefaultFactory)
+def notify_on_submit_message_default_factory(context):
+    return _(
+        "notify_on_submit_message_default_value",
+        "Booking ${booking_type} for ${booking_date} at ${booking_time} was created.<a href=${booking_print_url}>Link</a>",
+    )
+
+
+@provider(IContextAwareDefaultFactory)
+def notify_on_confirm_subject_default_factory(context):
+    return _(
+        "notify_on_confirm_subject_default_value",
+        "Booking of ${booking_date} at ${booking_time} was accepted",
+    )
+
+
+@provider(IContextAwareDefaultFactory)
+def notify_on_confirm_message_default_factory(context):
+    return _(
+        "notify_on_confirm_message_default_value",
+        "The booking${booking_type} for ${title} was confirmed! <a href=${booking_print_url}>Link</a>",
+    )
+
+
+@provider(IContextAwareDefaultFactory)
+def notify_on_move_subject_default_factory(context):
+    return _(
+        "notify_on_move_subject_default_value",
+        "Modified the boolking date for ${title}",
+    )
+
+
+@provider(IContextAwareDefaultFactory)
+def notify_on_move_message_default_factory(context):
+    return _(
+        "notify_on_move_message_default_value",
+        "The booking scheduling of ${booking_type} was modified."
+        "The new one is on ${booking_date} at ${booking_time}. <a href=${booking_print_url}>Link</a>.",
+    )
+
+
+@provider(IContextAwareDefaultFactory)
+def notify_on_refuse_subject_default_factory(context):
+    return _(
+        "notify_on_refuse_subject_default_value",
+        "Booking refused for ${title}",
+    )
+
+
+@provider(IContextAwareDefaultFactory)
+def notify_on_refuse_message_default_factory(context):
+    return _(
+        "notify_on_refuse_message_default_value",
+        "The booking ${booking_type} of ${booking_date} at ${booking_time} was refused.",
     )
 
 
@@ -513,27 +604,7 @@ class IPrenotazioniFolder(model.Schema):
     form.mode(templates_usage="display")
     templates_usage = RichText(
         title=_("templates_usage_label", "Templates usage"),
-        default="I testi e l’oggetto delle notifiche email possono essere configurate usando le seguenti variabili:"
-        "<ul>"
-        "<li>${title} - Titolo della prenotazione.</li>"
-        "<li>${booking_gate} - Sportello della prenotazione.</li>"
-        "<li>${booking_human_readable_start} - Data e ora prenotazione con formattazione standard.</li>"
-        "<li>${booking_date} - Data prenotazione.</li>"
-        "<li>${booking_end_date} - Data fine prenotazione.</li>"
-        "<li>${booking_time} - Orario di inizio prenotazione.</li>"
-        "<li>${booking_time_end} - Orario di fine prenotazione.</li>"
-        "<li>${booking_code} - Ticket identificativo della prenotazione da utilizzare per chiamare il cittadino allo sportello ad attesa ultimata.</li>"
-        "<li>${booking_type} - Tipologia prenotazione.</li>"
-        "<li>${booking_print_url} - Link di riepilogo prenotazione.</li>"
-        "<li>${booking_url_with_delete_token} - Link per cancellare la prenotazione.</li>"
-        "<li>${booking_user_phone} - Numero di telefono del cittadino.</li>"
-        "<li>${booking_user_email} - Email del cittadino.</li>"
-        "<li>${booking_office_contact_phone} - Telefono ufficio, se compilato.</li>"
-        "<li>${booking_office_contact_pec} - PEC ufficio, se compilata.</li>"
-        "<li>${booking_office_contact_fax} - Fax ufficio, se compilato.</li>"
-        "<li>${booking_how_to_get_to_office} - Informazioni su come raggiungere l’ufficio, se compilate.</li>"
-        "<li>${booking_office_complete_address} - Indirizzo completo dell’ufficio, se compilato.</li>"
-        "</ul>",
+        defaultFactory=templates_usage_default_factory,
         output_mime_type="text/html",
     )
     notify_on_submit_subject = schema.TextLine(
@@ -542,7 +613,7 @@ class IPrenotazioniFolder(model.Schema):
             default="Prenotazione created notification subject.",
         ),
         description=_("notify_on_submit_subject_help", default=""),
-        default="Prenotazione creata correttamente per ${title}",
+        defaultFactory=notify_on_submit_subject_default_factory,
         required=False,
     )
     notify_on_submit_message = RichText(
@@ -552,9 +623,7 @@ class IPrenotazioniFolder(model.Schema):
         ),
         output_mime_type="text/html",
         description=_("notify_on_submit_message_help", default=""),
-        default="La prenotazione ${booking_type} per il ${booking_date} alle ${booking_time} è stata creata."
-        "Riceverete una mail di conferma quando la prenotazione verrà confermata definitivamente."
-        "Se non hai salvato o stampato il promemoria, puoi visualizzarlo <a href=${booking_print_url}>questo link</a>",
+        defaultFactory=notify_on_submit_message_default_factory,
         required=False,
     )
     notify_on_confirm_subject = schema.TextLine(
@@ -563,7 +632,7 @@ class IPrenotazioniFolder(model.Schema):
             default="Prenotazione confirmed notification subject.",
         ),
         description=_("notify_on_confirm_subject_help", default=""),
-        default="Prenotazione del ${booking_date} alle ${booking_time} accettata",
+        defaultFactory=notify_on_confirm_subject_default_factory,
         required=False,
     )
     notify_on_confirm_message = RichText(
@@ -573,9 +642,7 @@ class IPrenotazioniFolder(model.Schema):
         ),
         output_mime_type="text/html",
         description=_("notify_on_confirm_message_help", default=""),
-        default="La prenotazione ${booking_type} per ${title} è stata confermata!"
-        "Se non hai salvato o stampato il promemoria, puoi visualizzarlo su <a href=${booking_print_url}>questo link</a>"
-        "Se desideri cancellare la prenotazione, accedi a <a href=${booking_print_url}>questo link</a>",
+        defaultFactory=notify_on_confirm_message_default_factory,
         required=False,
     )
     notify_on_move_subject = schema.TextLine(
@@ -584,7 +651,7 @@ class IPrenotazioniFolder(model.Schema):
             default="Prenotazione moved notification subject.",
         ),
         description=_("notify_on_move_subject_help", default=""),
-        default="Modifica data di prenotazione per ${title}",
+        defaultFactory=notify_on_move_subject_default_factory,
         required=False,
     )
     notify_on_move_message = RichText(
@@ -594,9 +661,7 @@ class IPrenotazioniFolder(model.Schema):
         ),
         output_mime_type="text/html",
         description=_("notify_on_move_message_help", default=""),
-        default="L'orario della sua prenotazione ${booking_type} è stata modificato."
-        "La nuova data è ${booking_date} alle ore ${booking_time}."
-        "Controlla o stampa il nuovo promemoria su  <a href=${booking_print_url}>questo link</a>.",
+        defaultFactory=notify_on_move_message_default_factory,
         required=False,
     )
     notify_on_refuse_subject = schema.TextLine(
@@ -605,7 +670,7 @@ class IPrenotazioniFolder(model.Schema):
             default="Prenotazione refused notification subject.",
         ),
         description=_("notify_on_refuse_subject_help", default=""),
-        default="Prenotazione rifiutata per ${title}",
+        defaultFactory=notify_on_refuse_subject_default_factory,
         required=False,
     )
     notify_on_refuse_message = RichText(
@@ -615,7 +680,7 @@ class IPrenotazioniFolder(model.Schema):
         ),
         output_mime_type="text/html",
         description=_("notify_on_refuse_message_help", default=""),
-        default="La prenotazione ${booking_type} del ${booking_date} delle ore ${booking_time} è stata rifiutata.",
+        defaultFactory=notify_on_refuse_message_default_factory,
         required=False,
     )