+++ CHANGES.rst
@@ -5,9 +5,9 @@ Changelog
 2.0.0.dev3 (unreleased)
 -----------------------
 
-- Handle contentrules by the plone events
+- Handle contentrules by the plone events and do not use contentrules anymore.
   [folix-01]
-- Change "day" type in week_table (TODO: need an upgrade step?)
+- Change "day" type in week_table (TODO: need an upgrade step?).
   [mauro]
 
 2.0.0.dev2 (2023-06-30)